+++ tests/test_compile_error.rs
@@ -3,4 +3,5 @@ fn test_compile_errors() {
     let t = trybuild::TestCases::new();
     t.compile_fail("tests/ui/reject_generics.rs");
     t.compile_fail("tests/ui/too_many_args_to_getter.rs");
+    t.compile_fail("tests/ui/invalid_pymethod_names.rs");
 }